pyproject.toml
[project]
name = "smart-cpa-bot"
version = "0.1.0"
description = "Dual Telegram bot system for CPA offer onboarding and offer delivery"
readme = "README.md"
authors = [{ name = "Smart CPA" }]
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "cachetools>=5.3",
    "tenacity>=8.2",
    "shortuuid>=1.0",
    "babel>=2.14"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "anyio>=4.2"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
